refactor(core): extract runPipeline layout, scene, contracts, candidates
Peel the remaining large runPipeline phases into focused modules:
panel layout, geometry/scene assembly, per-layer contracts, and
interaction candidate construction. The orchestrator is now a thin
pipeline of phase calls. Characterization tests cover backends,
tooltip fields, scaled constants, and facet layout placement.
